webapp/requirements.txt
Flask==3.0.2
google-cloud-storage==2.15.0
gunicorn==20.1.0